Split out of decimal_builder (#1843)
